Make pipeline compatible to add deep learning

+++ pyEDA/preprocessing.py
@@ -4,7 +4,7 @@ import scipy.signal as sps
 
 
 def resample_data(gsrdata, prevSR, newSR):
-	number_of_samples = round(len(gsrdata) * float(newSR) / prevSR)
+	number_of_samples = int(round(len(gsrdata) * float(newSR) / prevSR))
 	data = sps.resample(gsrdata, number_of_samples)
 	return data
